# Optimize the build time