This pull request attempts to optimize the build time, especially on Travis, by:

* letting `make' run multiple jobs in parallel
* starting with the sources that take longer to compile

These are the individual source compilation times, as measured on an AMD Ryzen 7 2700, sorted in decreasing order:

```text
43.485 s  mag_parser.cpp
 2.772 s  time_integration.cpp
 2.449 s  feellgoodSettings.cpp
 1.964 s  solver.cpp
 1.196 s  tetra.cpp
 0.972 s  facette.cpp
 0.941 s  read.cpp
 0.936 s  save.cpp
 0.770 s  linear_algebra.cpp
 0.715 s  femutil.cpp
 0.674 s  energy.cpp
 0.649 s  recentering.cpp
```

The cmake `SOURCES` variable has been sorted in this order, and it appears `make` follows the order of this variable.