@edgar-bonet 2019-10-19
Here are the build times reported by Travis for the current master branch and this pull request's head branch and merge commit:

|   commit   | dependencies | FeeLLGood | Total |
|------------|-------------:|----------:|------:|
| [master][] |     62.11    |   116.74  |  192  |
| [head][]   |     51.45    |    93.71  |  157  |
| [merge][]  |     60.19    |   107.34  |  181  |

“dependencies” is the time taken to download, build and install ANN, exprtk, ScalFMM and GMM. “FeeLLGood” is the time for building FeeLLGood, “total” is the total time of the test, including booting the virtual machine. I don't understand why “head” and “merge” are different: they have the exact same contents.

[master]: https://travis-ci.com/feellgood/FeeLLGood/builds/132659689
[head]: https://travis-ci.com/feellgood/FeeLLGood/builds/132668573
[merge]: https://travis-ci.com/feellgood/FeeLLGood/builds/132668700