@edgar-bonet 2019-11-01
Solved a conflict by rebasing on top of master.